관리자모드 메뉴분류관리 1차메뉴 추가

+++ src/main/java/com/example/junguniv_bb/domain/managermenu/model/ManagerMenu.java
@@ -29,7 +29,7 @@ public class ManagerMenu extends BaseTime {
     private String menuName; // 메뉴명
 
     @Column(name = "SORTNO")
-    private Integer sortno; // 정렬값
+    private Long sortno; // 정렬값
 
     @Column(name = "CHK_USE", length = 1)
     private String chkUse; // 사용여부
